Check for right type of boolean switches
Especially when passing these switches from the command line it is hard to get
the type right (you actually have to pass a JSON/YAML map to get this right,
see https://github.com/ansible/ansible/issues/17193), so explicitly test for it.
If the type is wrong any non-zero string will be interpreted as true in certain
contexts even when the string is "false".